refactor builderx to builder
