+++ core/stores/builder/builder_test.go
@@ -1,4 +1,4 @@
-package builderx
+package builder
 
 import (
 	"testing"
@@ -7,11 +7,11 @@ import (
 )
 
 type mockedUser struct {
-	ID string `db:"id" json:"id,omitempty"`
+	ID       string `db:"id" json:"id,omitempty"`
 	UserName string `db:"user_name" json:"userName,omitempty"`
-	Sex  int    `db:"sex" json:"sex,omitempty"`
-	UUID string `db:"uuid" uuid:"uuid,omitempty"`
-	Age  int    `db:"age" json:"age"`
+	Sex      int    `db:"sex" json:"sex,omitempty"`
+	UUID     string `db:"uuid" uuid:"uuid,omitempty"`
+	Age      int    `db:"age" json:"age"`
 }
 
 func TestFieldNames(t *testing.T) {
